Update common_gps.inc
Added an additional if check for the new T8 faction region system which adds a line to the clock to say which faction controls the region. Breaks findCoords() when that third line is present.

Added a check so that if coordinates are not found in the normal place (last line of the clock) it'll check the previous line to see if coords are there. The new line doesn't show up if no faction controls a region, so both cases need to be handled.