Update function drawWater - See comments
When using statusScreen  and 'if not drawWater()`  statement, stop the "Refilling water" message, because it's not true.

Don't overRide a statusScreen with a  sleepStatus message (and any new buttons) from disappearing while checking/attemptingto drawWater

In other words, don't display a "Refilling water" message if not water source was found.